pyproject.toml
[tool.poetry]
name = "coreason_adlc_api"
version = "0.2.0"
description = "Secure ADLC Middleware enforcing PII scrubbing, budget caps, and strict governance."
authors = ["Gowtham A Rao <gowtham.rao@coreason.ai>"]
license = "Proprietary"
readme = "README.md"
packages = [{include = "coreason_adlc_api", from = "src"}]

[tool.poetry.scripts]
coreason-api = "coreason_adlc_api.main:main"

[tool.poetry.dependencies]
python = ">=3.12, <3.15"
loguru = "^0.7.2"
fastapi = "^0.127.0"
uvicorn = {extras = ["standard"], version = "^0.40.0"}
pydantic-settings = "^2.12.0"
cryptography = "^46.0.3"
asyncpg = "^0.31.0"
pyjwt = {extras = ["crypto"], version = "^2.10.1"}
email-validator = "^2.3.0"
litellm = "^1.80.11"
presidio-analyzer = {version = "^2.2.360", python = "<3.14"}
spacy = "^3.8.11"
redis = "^7.1.0"
httpx = "^0.28.1"
keyring = "^24.0.0"
alembic = "^1.17.2"
sqlalchemy = "^2.0.45"
psycopg2-binary = "^2.9.11"

[tool.poetry.group.dev.dependencies]
types-redis = "^4.6.0.20241004"
pytest = "^8.2.2"
ruff = "^0.4.8"
pre-commit = "^3.7.1"
pytest-cov = "^5.0.0"
mkdocs = "^1.6.0"
mkdocs-material = "^9.5.26"
pytest-asyncio = "^1.3.0"
mypy = "^1.19.1"
types-pyyaml = "^6.0.12.20250915"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[project]
name = "coreason_adlc_api"
version = "0.2.0"
description = "Secure ADLC Middleware enforcing PII scrubbing, budget caps, and strict governance."
readme = "README.md"
requires-python = ">=3.12, <3.15"
authors = [
    { name = "Gowtham A Rao", email = "gowtham.rao@coreason.ai" },
]
license = { file = "LICENSE" }
classifiers = [
    "License :: Other/Proprietary License",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Operating System :: OS Independent",
]

[project.urls]
Homepage = "https://github.com/CoReason-AI/coreason_adlc_api"
Repository = "https://github.com/CoReason-AI/coreason_adlc_api"
Documentation = "https://github.com/CoReason-AI/coreason_adlc_api"

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "B", "I"]
ignore = ["B008"]

[tool.ruff.lint.isort]
known-first-party = ["coreason_adlc_api"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = "--cov=src --cov-report=term-missing --cov-fail-under=100"
testpaths = ["tests"]

[tool.coverage.run]
omit = ["tests/*"]
